runtime: address playback review feedback
