+++ docs/maintainers/speakswiftly-api-coverage-matrix.md
@@ -69,8 +69,7 @@ That means the server is best understood as a transport adapter over the public 
 | `runtime.generate.snapshot()` | Full | `GET /generation/queue` | `speak-swiftly://overview` | Exposed directly from runtime-owned generation queue data. |
 | `runtime.jobs.list()` / `job(id:)` / `expire(id:)` | Full | `GET /generation/jobs`, `GET /generation/jobs/{job_id}`, `DELETE` equivalent via expiry route family when present in HTTP flow | `expire_generation_job`, `speak-swiftly://generation/jobs`, `speak-swiftly://generation/jobs/{job_id}` | Retained generation-job reads use resources; expiry remains a tool because it mutates retained runtime state. |
 | `runtime.artifacts()` / `artifact(id:)` | Full | `GET /generation/artifacts`, `GET /generation/artifacts/{artifact_id}` | `speak-swiftly://generation/artifacts`, `speak-swiftly://generation/artifacts/{artifact_id}` | Saved artifact reads use one retained artifact family; batch membership is read through generation jobs instead of a separate batch read family. |
-| `runtime.playback.snapshot()` | Full | `GET /playback/queue` | `speak-swiftly://overview`, `speak-swiftly://playback/queue` | Exposed as the playback queue read model. |
-| `runtime.playback.snapshot()` | Full | `GET /playback/state` | `speak-swiftly://overview`, `speak-swiftly://playback` | Used directly for playback state reads and control settling. |
+| `runtime.playback.snapshot()` | Full | `GET /playback/queue`, `GET /playback/state` | `speak-swiftly://overview`, `speak-swiftly://playback/queue`, `speak-swiftly://playback` | Exposed as both the playback queue read model and the playback state read/control-settling source. |
 | `runtime.playback.pause()` / `resume()` | Full | `POST /playback/pause`, `POST /playback/resume` | `pause_playback`, `resume_playback` | The server now aligns its cached playback snapshot with these accepted control responses. |
 | `runtime.playback.clearQueue()` | Full | `DELETE /playback/queue` | `clear_playback_queue` | Returns cleared queued-count information rather than forcing clients to infer it. |
 | `runtime.cancelRequest(_:)` / queue-scoped cancellation | Full | `DELETE /requests/{request_id}` with optional `?scope=generation\|playback` | `cancel_request` with optional `scope` | Cancels one active or queued request by id, with scope only when the caller deliberately needs queue-specific protection. |
